feat: implement custom hook for blog posts and refactor HomePage and ViewPostPage to use it
